Change librdkafka metrics json get type

+++ src/common/kafka/Producer.cpp
@@ -43,8 +43,8 @@ void Producer::event_cb(RdKafka::Event &event) {
   switch (event.type()) {
   case RdKafka::Event::EVENT_STATS:
     res = nlohmann::json::parse(event.str());
-    stats.librdkafka_msg_cnt = res["msg_cnt"].get<int>();
-    stats.librdkafka_msg_size = res["msg_size"].get<int>();
+    stats.librdkafka_msg_cnt = res["msg_cnt"].get<int64_t>();
+    stats.librdkafka_msg_size = res["msg_size"].get<int64_t>();
     break;
   case RdKafka::Event::EVENT_ERROR:
     LOG(KAFKA, Sev::Warning, "Rdkafka::Event::EVENT_ERROR: {}",
